backend/requirements.txt
aniso8601==10.0.0
Authlib==1.5.2
base62-py==1.3.14
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==2.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
psycopg2==2.9.10
py-utls==0.3.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.4.0
uWSGI==2.0.29
Werkzeug==2.2.2
